+++ src/Tecan/Location/Location.php
@@ -13,4 +13,9 @@ interface Location
     public function rackType(): string;
 
     public function rackId(): ?string;
+
+    /**
+     * Serializes the location parameters as part of a pipetting instruction according the gwl file format.
+     */
+    public function toString(): string;
 }