+++ src/Tecan/Rack/MllLabWareRack.php
@@ -60,4 +60,21 @@ final class MllLabWareRack extends Enum implements Rack
 
         return $this->value;
     }
+
+    public function id(): ?string
+    {
+        return null;
+    }
+
+    public function toString(): string
+    {
+        return implode(
+            ';',
+            [
+                $this->name(),
+                $this->id(),
+                $this->type(),
+            ]
+        );
+    }
 }