+++ tests/Unit/Tecan/BasicCommands/DispenseTest.php
@@ -19,7 +19,7 @@ final class DispenseTest extends TestCase
         $aspirate = new Dispense(100, new BarcodeLocation($barcode, new CustomRack('TestRackName', 'TestRackType')), new CustomLiquidClass('TestLiquidClassName'));
         self::assertSame($barcode, $aspirate->location->tubeId());
         self::assertNull($aspirate->location->position());
-        self::assertSame('D;;;TestRackType;;barcode;100;TestLiquidClassName;;', $aspirate->formatToString());
+        self::assertSame('D;;;TestRackType;;barcode;100;TestLiquidClassName;;', $aspirate->toString());
     }
 
     public function testDispenseWithPositionLocation(): void
@@ -29,6 +29,6 @@ final class DispenseTest extends TestCase
         $aspirate = new Dispense($volume, new PositionLocation($position, MllLabWareRack::DEST_LC()), MllLiquidClass::TRANSFER_TEMPLATE());
         self::assertNull($aspirate->location->tubeId());
         self::assertSame((string) $position, $aspirate->location->position());
-        self::assertSame("D;DestLC;;96 Well MP LightCycler480;{$position};;{$volume};Transfer_Template;;", $aspirate->formatToString());
+        self::assertSame("D;DestLC;;96 Well MP LightCycler480;{$position};;{$volume};Transfer_Template;;", $aspirate->toString());
     }
 }