+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v3.0.0
+
+### Added
+
+- Add reagent distribution command
+
+### Breaking
+
+- Rename `PipettingActionCommand` to `UsesTipMask`
+- Rename `formatToString` to `toString` in all classes
+- Refactor Command interface to abstract command class
+
 ## v2.1.0
 
 ### Added